+++ tests/_legacy/test_menu_creation.py
@@ -2,7 +2,7 @@ import os
 import sys
 
 import pytest
-from conda.cli.python_api import run_command
+from conda.testing.fixtures import conda_cli
 
 if sys.platform == "win32":
     import menuinst._legacy as menuinst
@@ -48,24 +48,24 @@ class TestWindowsShortcuts(object):
         open(nonadmin, 'a').close()
         shortcut = os.path.join(menu_dir, "menu-windows.json")
         test_env_name = 'test_env'
-        run_command("create", "-n", test_env_name, "python")
+        conda_cli("create", "-n", test_env_name, "python")
         prefix = os.path.join(sys.prefix, 'envs', test_env_name)
         name = 'Anaconda Prompt (%s)' % test_env_name
         menuinst.install(shortcut, prefix=prefix, remove=False)
         assert file_exist('user', name)
         menuinst.install(shortcut, prefix=prefix, remove=True)
         assert not file_exist('user', name)
-        run_command("remove", "-n", test_env_name, "--all")
+        conda_cli("remove", "-n", test_env_name, "--all")
 
     def test_root_prefix(self):
         nonadmin = os.path.join(sys.prefix, ".nonadmin")
         open(nonadmin, 'a').close()
         shortcut = os.path.join(menu_dir, "menu-windows.json")
         root_prefix = os.path.join(menu_dir, 'temp_env')
-        run_command("create", "-p", root_prefix, "python")
+        conda_cli("create", "-p", root_prefix, "python")
         name = 'Anaconda Prompt (%s)' % os.path.split(sys.prefix)[1]
         menuinst.install(shortcut, remove=False, root_prefix=root_prefix)
         assert file_exist('user', name)
         menuinst.install(shortcut, remove=True, root_prefix=root_prefix)
         assert not file_exist('user', name)
-        run_command("remove", "-p", root_prefix, "--all")
+        conda_cli("remove", "-p", root_prefix, "--all")